refactor: Remove unnecessary duplicate types from `Types.ts`
In fact, some of these types no longer even exist in the backend / primer-api.
